+++ ecoz/src/ecoz/hmm/HmmClassifier.scala
@@ -1,7 +1,7 @@
 package ecoz.hmm
 
-import ecoz.rpt.ConfusionMatrix
-import ecoz.rpt.quoted
+import ecoz.rpt
+import ecoz.rpt.{ConfusionMatrix, quoted}
 import ecoz.symbol.SymbolSequence
 
 class HmmClassifier(hmms: List[Hmm],
@@ -11,7 +11,7 @@ class HmmClassifier(hmms: List[Hmm],
   private val classNames = collection.mutable.HashSet[String]()
   classNames ++= hmms.map(_.classNameOpt.get)
 
-  private val bySeqName = collection.mutable.HashMap[String, List[List[(Hmm,BigDecimal)]]]()
+  private val bySeqName = collection.mutable.HashMap[String, List[(SymbolSequence, List[(Hmm,BigDecimal)])]]()
 
   /**
     * Performs classification of the given sequence,
@@ -25,24 +25,33 @@ class HmmClassifier(hmms: List[Hmm],
     val list = bySeqName.getOrElseUpdate(seqName, List.empty)
     val hmmsAndProbs = hmms map { hmm ⇒ (hmm, hmm.probability(seq)) }
     val sortedByProb = hmmsAndProbs.sortBy(-_._2)
-    bySeqName.update(seqName, sortedByProb :: list)
-    sortedByProb
+    bySeqName.update(seqName, (seq, sortedByProb) :: list)
+    sortedByProb map (e ⇒ (e._1, e._2))
   }
 
   def reportResults(): Unit = {
     val cm = new ConfusionMatrix(classNames.toSet)
 
     bySeqName foreach { case (seqName, instances) ⇒
-      instances foreach { sortedByProb ⇒
+      instances foreach { case (seq, sortedByProb) ⇒
         val hmmWinner = sortedByProb.head._1
-        val hmmName = hmmWinner.classNameOpt.get
-        cm.setWinner(seqName, hmmName)
+        val hmmWinnerName = hmmWinner.classNameOpt.get
+        cm.setWinner(seqName, hmmWinnerName)
 
-        if (seqName != hmmName && showRanked > 0) {
-          printf(s"  %24s sequence:\n", quoted(seqName))
-          sortedByProb.take(showRanked) foreach { case (hmm, prob) ⇒
-            val pStr = "%s" format quoted(hmm.classNameOpt.get)
-            printf(s"  %24s prob = %s\n", pStr, prob)
+        if (seqName != hmmWinnerName && showRanked > 0) {
+          println()
+          printf(s"  %-24s sequence (T=%d):\n", quoted(seqName), seq.T)
+          sortedByProb.zipWithIndex.take(showRanked) foreach { case ((hmm, prob), rank) ⇒
+            val hmmNameStr = "%s" format quoted(hmm.classNameOpt.get)
+
+            val (asterisk, probStr) = if (hmm.classNameOpt.contains(seqName))
+              ("*", rpt.magenta(prob).toString())
+            else
+              (" ", prob.toString())
+
+            printf(s"  %s [%02d] %-24s p = %s\n",
+              asterisk, rank, hmmNameStr, probStr
+            )
           }
           println()
         }